+++ src/com/simplequeries/SQQuery/UpdateQuery.java
@@ -52,8 +52,8 @@ public class UpdateQuery {
      *      ltOrEq: less than or equals ( <= )
      *      gtOrEq: greater than or equals ( >= )
      */
-    public UpdateQuery eq(String arg) { return insertComparison(" = ", arg); }
-    public UpdateQuery notEq(String arg) { return insertComparison(" <!= ", arg); }
+    public UpdateQuery eq(Object arg) { return insertComparison(" = ", arg); }
+    public UpdateQuery notEq(Object arg) { return insertComparison(" <!= ", arg); }
     public UpdateQuery gt(Object arg) { return insertComparison(" > ", arg); }
     public UpdateQuery lt(Object arg) { return insertComparison(" < ", arg); }
     public UpdateQuery gtOrEq(Object arg) { return insertComparison(" >= ", arg); }
@@ -95,12 +95,23 @@ public class UpdateQuery {
      * @return the updated query object
      */
     public UpdateQuery in(String... args) {
-        query.append(" IN (");
-        for (String arg : args) {
-            query.append("\'").append(arg).append("\'").append(", ");
+        if (args.length == 1) {
+            for (String arg : args) {
+                if (arg.startsWith("(")) { // nested select
+                    query.append(" IN ").append(arg);
+                } else { // just single value, same as below
+                    query.append(" IN (").append("\'").append(arg).append("\'").append(")");
+                }
+            }
+            return this;
+        } else {
+            query.append(" IN (");
+            for (String arg : args) {
+                query.append("\'").append(arg).append("\'").append(", ");
+            }
+            query.delete(query.lastIndexOf(", "), query.length()).append(")");
+            return this;
         }
-        query.delete(query.lastIndexOf(", "), query.length()).append(")");
-        return this;
     }
 
     /**
@@ -126,6 +137,16 @@ public class UpdateQuery {
     public UpdateQuery between(Object least, Object most) {
         query.append(" BETWEEN ").append(least);
         and(most.toString());
+        return this;
+    }
+
+    /**
+     * NOT clause.
+     *
+     * @return the updated query object
+     */
+    public UpdateQuery not() {
+        query.append(" NOT");
         return this;
     }
